+++ pol-core/pol/packetscrobj.h
@@ -33,7 +33,7 @@ public:
     virtual std::string getStringRep() const;
     bool SetSize(u16 newsize);
     BObjectImp* SetSize(u16 newsize, bool giveReturn);
-    virtual unsigned int sizeEstimate() const { return buffer.size(); }
+    virtual size_t sizeEstimate() const { return buffer.size(); }
     virtual const char* typeOf() const { return "Packet"; }
 	virtual int typeOfInt() const { return OTPacket; }
